feat(assistant): add full-page governance editor shell
