+++ packages/desktop/src/renderer/services/i18n/i18n-keys.d.ts
@@ -1169,6 +1169,7 @@ export type I18nKey =
   | 'settings.assistantCreated'
   | 'settings.assistantDescription'
   | 'settings.assistantDescriptionPlaceholder'
+  | 'settings.assistantExtensionReadonlyTip'
   | 'settings.assistantFilterAll'
   | 'settings.assistantFilterBuiltin'
   | 'settings.assistantFilterCustom'
@@ -1426,6 +1427,8 @@ export type I18nKey =
   | 'settings.envLabel'
   | 'settings.envValuePlaceholder'
   | 'settings.extensionAssistantReadonly'
+  | 'settings.extensionSkills'
+  | 'settings.extensionSkillsBadge'
   | 'settings.feedback'
   | 'settings.fontSizeChat'
   | 'settings.fontSizeCode'
