+++ packages/desktop/src/renderer/services/i18n/locales/en-US/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "Add Skills",
   "customSkills": "Custom Skills",
   "builtinSkills": "Builtin Skills",
+  "extensionSkills": "Extension Skills",
   "noCustomSkills": "No custom skills added",
   "noBuiltinSkills": "No builtin skills available",
+  "extensionSkillsBadge": "Extension",
   "autoInjectedSkills": "Auto-injected Skills",
   "autoInjectedSkillsBadge": "Auto",
+  "assistantExtensionReadonlyTip": "Extension assistants are read-only in assistant settings.",
   "skillFolderPath": "Skill Folder Path",
   "skillPathPlaceholder": "Enter or browse skill folder path(s), separated by comma",
   "pleaseSelectSkillPath": "Please select a skill folder path",
