+++ packages/desktop/src/renderer/services/i18n/locales/ja-JP/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "スキルを追加",
   "customSkills": "カスタムスキル",
   "builtinSkills": "ビルトインスキル",
+  "extensionSkills": "拡張スキル",
   "noCustomSkills": "カスタムスキルはありません",
   "noBuiltinSkills": "利用可能なビルトインスキルはありません",
+  "extensionSkillsBadge": "拡張",
   "autoInjectedSkills": "自動注入スキル",
   "autoInjectedSkillsBadge": "自動",
+  "assistantExtensionReadonlyTip": "拡張アシスタントはアシスタント設定では読み取り専用です。",
   "skillFolderPath": "スキルフォルダパス",
   "skillPathPlaceholder": "スキルフォルダパスを入力または参照",
   "pleaseSelectSkillPath": "スキルフォルダパスを選択してください",
