+++ packages/desktop/src/renderer/services/i18n/locales/ko-KR/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "스킬 추가",
   "customSkills": "사용자 정의 스킬",
   "builtinSkills": "내장 스킬",
+  "extensionSkills": "확장 스킬",
   "noCustomSkills": "사용자 정의 스킬 없음",
   "noBuiltinSkills": "사용 가능한 내장 스킬 없음",
+  "extensionSkillsBadge": "확장",
   "autoInjectedSkills": "자동 주입 스킬",
   "autoInjectedSkillsBadge": "자동",
+  "assistantExtensionReadonlyTip": "확장 어시스턴트는 어시스턴트 설정에서 읽기 전용입니다.",
   "skillFolderPath": "스킬 폴더 경로",
   "skillPathPlaceholder": "스킬 폴더 경로 입력 또는 찾아보기",
   "pleaseSelectSkillPath": "스킬 폴더 경로를 선택하세요",
