+++ packages/desktop/src/renderer/services/i18n/locales/pt-BR/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "Adicionar Habilidades",
   "customSkills": "Habilidades Personalizadas",
   "builtinSkills": "Habilidades Integradas",
+  "extensionSkills": "Habilidades de Extensão",
   "noCustomSkills": "Nenhuma habilidade personalizada adicionada",
   "noBuiltinSkills": "Nenhuma habilidade integrada disponível",
+  "extensionSkillsBadge": "Extensão",
   "autoInjectedSkills": "Habilidades Injetadas Automaticamente",
   "autoInjectedSkillsBadge": "Auto",
+  "assistantExtensionReadonlyTip": "Assistentes de extensão são somente leitura nas configurações de assistente.",
   "skillFolderPath": "Caminho da Pasta da Habilidade",
   "skillPathPlaceholder": "Insira ou navegue pelos caminhos da pasta da habilidade, separados por vírgula",
   "pleaseSelectSkillPath": "Por favor, selecione o caminho de uma pasta de habilidade",
