+++ packages/desktop/src/renderer/services/i18n/locales/ru-RU/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "Добавить навыки",
   "customSkills": "Пользовательские навыки",
   "builtinSkills": "Встроенные навыки",
+  "extensionSkills": "Навыки расширений",
   "noCustomSkills": "Пользовательские навыки не добавлены",
   "noBuiltinSkills": "Встроенные навыки недоступны",
+  "extensionSkillsBadge": "Расширение",
   "autoInjectedSkills": "Автоматически добавленные навыки",
   "autoInjectedSkillsBadge": "Авто",
+  "assistantExtensionReadonlyTip": "Ассистенты расширений доступны только для чтения в настройках ассистентов.",
   "skillFolderPath": "Путь к папке навыков",
   "skillPathPlaceholder": "Введите или выберите пути к папкам навыков, разделённые запятыми",
   "pleaseSelectSkillPath": "Выберите путь к папке навыков",
