+++ packages/desktop/src/renderer/services/i18n/locales/tr-TR/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "Yetenek Ekle",
   "customSkills": "Özel Yetenekler",
   "builtinSkills": "Yerleşik Yetenekler",
+  "extensionSkills": "Uzantı Yetenekleri",
   "noCustomSkills": "Eklenmiş özel yetenek yok",
   "noBuiltinSkills": "Mevcut yerleşik yetenek yok",
+  "extensionSkillsBadge": "Uzantı",
   "autoInjectedSkills": "Otomatik Eklenen Beceriler",
   "autoInjectedSkillsBadge": "Otomatik",
+  "assistantExtensionReadonlyTip": "Uzantı asistanları asistan ayarlarında salt okunurdur.",
   "skillFolderPath": "Yetenek Klasör Yolu",
   "skillPathPlaceholder": "Yetenek klasör yolunu girin veya göz atın (virgülle ayırın)",
   "pleaseSelectSkillPath": "Lütfen bir yetenek klasör yolu seçin",
