+++ packages/desktop/src/renderer/services/i18n/locales/uk-UA/settings.json
@@ -97,8 +97,13 @@
   "addSkills": "Додати навички",
   "customSkills": "Власні навички",
   "builtinSkills": "Вбудовані навички",
+  "extensionSkills": "Навички розширень",
   "noCustomSkills": "Власних навичок не додано",
   "noBuiltinSkills": "Вбудованих навичок немає",
+  "extensionSkillsBadge": "Розширення",
+  "autoInjectedSkills": "Автоматично додані навички",
+  "autoInjectedSkillsBadge": "Авто",
+  "assistantExtensionReadonlyTip": "Асистенти розширень доступні лише для читання в налаштуваннях асистентів.",
   "skillFolderPath": "Шлях до папки з навичками",
   "skillPathPlaceholder": "Введіть або виберіть шлях(и) до папок з навичками, розділені комами",
   "pleaseSelectSkillPath": "Будь ласка, виберіть шлях до папки з навичками",
