+++ packages/desktop/src/renderer/services/i18n/locales/zh-CN/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "添加技能",
   "customSkills": "自定义技能",
   "builtinSkills": "内置技能",
+  "extensionSkills": "扩展技能",
   "noCustomSkills": "暂无自定义技能",
   "noBuiltinSkills": "暂无可用的内置技能",
+  "extensionSkillsBadge": "扩展",
   "autoInjectedSkills": "自动注入技能",
   "autoInjectedSkillsBadge": "自动",
+  "assistantExtensionReadonlyTip": "扩展助手在助手设置中为只读。",
   "skillFolderPath": "技能文件夹路径",
   "skillPathPlaceholder": "输入或浏览技能文件夹路径，多个路径请用逗号分隔",
   "pleaseSelectSkillPath": "请选择技能文件夹路径",
