+++ packages/desktop/src/renderer/services/i18n/locales/zh-TW/settings.json
@@ -97,10 +97,13 @@
   "addSkills": "新增技能",
   "customSkills": "自訂技能",
   "builtinSkills": "內建技能",
+  "extensionSkills": "擴充技能",
   "noCustomSkills": "尚無自訂技能",
   "noBuiltinSkills": "尚無可用的內建技能",
+  "extensionSkillsBadge": "擴充",
   "autoInjectedSkills": "自動注入技能",
   "autoInjectedSkillsBadge": "自動",
+  "assistantExtensionReadonlyTip": "擴充助理在助手設定中為唯讀。",
   "skillFolderPath": "技能資料夾路徑",
   "skillPathPlaceholder": "輸入或瀏覽技能資料夾路徑",
   "pleaseSelectSkillPath": "請選擇技能資料夾路徑",
